refactor: move resource provider into android-utils
