+++ features/resume/internal/src/main/kotlin/uk/gov/onelogin/criorchestrator/features/resume/internal/AndroidResourceProvider.kt
@@ -4,14 +4,9 @@ import android.content.Context
 import androidx.annotation.StringRes
 import com.squareup.anvil.annotations.ContributesBinding
 import uk.gov.logging.api.analytics.extensions.getEnglishString
+import uk.gov.onelogin.criorchestrator.libraries.androidutils.ResourceProvider
 import uk.gov.onelogin.criorchestrator.libraries.di.CriOrchestratorScope
 import javax.inject.Inject
-
-fun interface ResourceProvider {
-    fun getEnglishString(
-        @StringRes resId: Int,
-    ): String
-}
 
 @ContributesBinding(CriOrchestratorScope::class)
 class AndroidResourceProvider
