+++ libraries/android-utils/src/androidTest/kotlin/uk/gov/onelogin/criorchestrator/libraries/androidutils/EmptyInstrumentationTest.kt
@@ -1,0 +1,12 @@
+package uk.gov.onelogin.criorchestrator.libraries.androidutils
+
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import org.junit.Test
+import org.junit.runner.RunWith
+
+@RunWith(AndroidJUnit4::class)
+class EmptyInstrumentationTest {
+    @Test
+    fun emptyTest() {
+    }
+}
