+++ libraries/android-utils/src/main/kotlin/uk/gov/onelogin/criorchestrator/libraries/androidutils/ResourceProvider.kt
@@ -1,0 +1,9 @@
+package uk.gov.onelogin.criorchestrator.libraries.androidutils
+
+import androidx.annotation.StringRes
+
+fun interface ResourceProvider {
+    fun getEnglishString(
+        @StringRes resId: Int,
+    ): String
+}
