+++ libraries/android-utils/src/test/kotlin/uk/gov/onelogin/criorchestrator/libraries/androidutils/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
 package uk.gov.onelogin.criorchestrator.libraries.androidutils
 
-import org.junit.Test
-
-import org.junit.Assert.*
+import org.junit.jupiter.api.Assertions.assertEquals
+import org.junit.jupiter.api.Test
 
 /**
  * Example local unit test, which will execute on the development machine (host).
