+++ libraries/android-utils/src/testFixtures/kotlin/uk/gov/onelogin/criorchestrator/libraries/androidutils/FakeResourceProvider.kt
@@ -1,0 +1,5 @@
+package uk.gov.onelogin.criorchestrator.libraries.androidutils
+
+class FakeResourceProvider : ResourceProvider {
+    override fun getEnglishString(resId: Int): String = "dummy string"
+}
